[Fix]: user 데이터 추가 로직 작성  및 user/ 하위 폴더 구조 변경
